requirements.txt
numpy<2
torch>=2.1.0
torchvision>=0.16.0
roboflow>=1.0.5
supervision>=0.18.0
albumentations==1.3.0
Pillow>=10.0.0
tqdm>=4.66.1
opencv-python>=4.8.0
scikit-learn>=1.3.0
Flask>=2.3.0
Werkzeug>=2.3.0